requirements.txt
netifaces==0.10.9
python-nmap==0.6.1
scapy==2.4.3